pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "creativity-agentscope"
version = "0.1.0"
description = "AgentScope creativity project for intelligent agent workflows"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "agentscope",
    "agentscope-runtime", 
    "python-dotenv",
    "dashscope>=1.24.4",
    "requests>=2.25.1",
    "psutil>=5.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["utils*", "workflow*", "tools*", "agents*", "apps*"]
exclude = ["tests*", "__pycache__*", "*.pyc"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt", "*.md"]

[project.scripts]
creativity-agentscope = "app:main"

[tool.black]
line-length = 200
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"